refactor: catalog renamed, added new tests
