+++ src/test/java/io/github/tomaszziola/javabuildautomaton/models/ProjectModel.java
@@ -1,7 +1,8 @@
 package io.github.tomaszziola.javabuildautomaton.models;
 
-import static io.github.tomaszziola.javabuildautomaton.build.BuildTool.GRADLE;
+import static io.github.tomaszziola.javabuildautomaton.buildsystem.BuildTool.GRADLE;
 
+import io.github.tomaszziola.javabuildautomaton.buildsystem.BuildTool;
 import io.github.tomaszziola.javabuildautomaton.project.Project;
 
 public final class ProjectModel {
@@ -9,11 +10,10 @@ public final class ProjectModel {
   private ProjectModel() {}
 
   public static Project basic() {
-    return new Project(
-        1L,
-        "test-project-from-db",
-        "TomaszZiola/test",
-        "/Users/Tomasz/Documents/IdeaProjects/test",
-        GRADLE);
+    return basic(GRADLE, "/Users/Tomasz/Documents/IdeaProjects/test");
+  }
+
+  public static Project basic(final BuildTool buildTool, final String localPath) {
+    return new Project(1L, "test-project-from-db", "TomaszZiola/test", localPath, buildTool);
   }
 }
